% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.birth_in_hospital.R
\name{birth_in_hospital}
\alias{birth_in_hospital}
\title{Birth In Hospital}
\value{
a \code{\link{logical}} column
}
\description{
A logical field indicating whether the birth occurred within a hospital.
}
\section{Data Quality Tests}{


This column is tested for the following quality assumptions prior to packaging:
\enumerate{
  \item non-NA values exist in each expected year
}
}

\seealso{
\code{\link{births}}

Other births-column: \code{\link{birth_month_date}},
  \code{\link{birth_state}},
  \code{\link{birth_weekday_date}}
}
