% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.birth_state.R
\name{birth_state}
\alias{birth_state}
\title{Birth State}
\value{
a \code{\link{factor}} column
}
\description{
A factor column describing the state or territory where the birth occurred.
}
\details{
From 1968 to 2004 the public data sets included the state (or territory) where the birth
occurred. After 2004, state of occurrence is no longer included, presumably for privacy reasons.
This field include all 50 states, and the District of Columbia (i.e. Washington D.C.).

The field includes placeholder levels for Puerto Rico, the Virgin Islands, and Guam, but due to
the way that the raw data are processed, there are no records for these territories. In future
development, this may be changed so that they are included.
}
\section{Data Quality Tests}{


This column is tested for the following quality assumptions prior to packaging:
\enumerate{
  \item all 51 states (including D.C.) are represented in expected years
  \item there are no missing values in expected years
  \item no states are represented after 2004
  \item territories have no records at any time
}
}

\seealso{
\code{\link{births}}

Other births-column: \code{\link{birth_in_hospital}},
  \code{\link{birth_month_date}},
  \code{\link{birth_via_cesarean}},
  \code{\link{birth_weekday_date}},
  \code{\link{child_sex}}, \code{\link{mother_age}}
}
