% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.child_sex.R
\name{child_sex}
\alias{child_sex}
\title{Child Sex}
\value{
a \code{\link{factor}} column
}
\description{
A factor field which represents the sex of the child.
}
\details{
A two-level factor which indicates whether the child was designated male or female at birth.
}
\section{Data Quality Tests}{


This column is tested for the following quality assumptions prior to packaging:
\enumerate{
  \item no missing data in field
}
}

\seealso{
\code{\link{births}}

Other births-column: \code{\link{birth_in_hospital}},
  \code{\link{birth_month_date}},
  \code{\link{birth_state}},
  \code{\link{birth_via_cesarean}},
  \code{\link{birth_weekday_date}},
  \code{\link{mother_age}}
}
