% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_evaluations.R
\name{cmaw}
\alias{cmaw}
\title{Calendar Month Artifact Weight}
\usage{
cmaw(month_date)
}
\arguments{
\item{month_date}{- A date that represents the month which a value is aggregated to}
}
\value{
A weighting value which can be used to smooth the noise caused by the Julian calendar
}
\description{
The purpose of this function is to provide a weight which can be used to adjust values that are
aggregated at the monthly level so that the noise caused by unequal intervals (i.e varying number
of days in each month) is muted.
}
\details{
Divide the equal interval month value (1/12) by the actual share of the month in the year that
The month and year for this calculation are obtained by the provided date variable. 
This calculation accounts for leap (366 day) years when they occur.
}
\examples{
 # determine the appropriate weighting for volumes in February 2016

 x = cmaw(as.Date('2016-02-01')
 print(x)
 1.048851 
 # then apply this weighting to volume values to get a scaled value

 32000 * x
 33563.22

}

