% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Vital Statistics Data Dictionary}
\usage{
data_dictionary()
}
\value{
A data dictionary of the attributes that are available from a reduction
of the various years of data publised by the CDC is their vital statistics data sets.
}
\description{
Vital Statistics Data Dictionary
}

