% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_extenders.R
\name{ext.birth_weekday}
\alias{ext.birth_weekday}
\alias{ext.birth_month}
\alias{ext.birth_year}
\alias{ext.birth_decade}
\title{Birth Data Field Extender}
\usage{
ext.birth_weekday(births_data)

ext.birth_month(births_data)

ext.birth_year(births_data)

ext.birth_decade(births_data)
}
\arguments{
\item{births_data}{this raw births data set, or a transformation of it (but be aware that any
changes may alter the results of these calculations)}
}
\value{
a data.table with the newly applied calculated field
}
\description{
The births data set has numerous useful attributes that can be derived from the data that is
included in the binary. However, because this data set is already very large, in the name of
saving storage space for the package, and memory when loading this data set for analysis, only
the bare minimum of information that is necessary to derive other useful columns is included.
}
\details{
As a convenience for accessing these columns, the package includes formulas to derive them. These
formulas only take one argument, as they are not meant to be customized to any degree. One
important thing to note is that these extensions will only work if you've created a copy of the
births dataset. This is best done using the \code{\link[data.table]{copy}} function on the births
table.
}
\examples{
dt = data.table::copy(births)
ext.birth_weekday(dt)
dt

}
