% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HHS_cesarean_1996}
\alias{HHS_cesarean_1996}
\title{Cesarean rates by age: United States, 1988-92}
\format{An object of class \code{\link{data.frame}}}
\source{
\href{http://www.cdc.gov/nchs/data/series/sr_13/sr13_124.pdf}{Trends in Hospital
  Utilization: United States, 1988-92}

  Table 26. Vital and Health Statistics, Series 13, Number 124, June 1996

  Gillum BS, Graves EJ, Kozak LJ. National Center for Health Statistics.
}
\usage{
HHS_cesarean_1996
}
\description{
A table of cesarean section rates between 1980 and 1992, with a breakdown by maternal age. These
rates are calculated with data collected via the National  Hospital Discharge Survey.
}
\keyword{datasets}
