% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{rate_lg}
\alias{rate_lg}
\title{Logical Field Rate Calculation}
\usage{
rate_lg(field, cases = 1)
}
\arguments{
\item{field}{- A name of a logical table field name that will be aggregated}

\item{cases}{- A value that should be applied to the result in order to simulate
a number of records with this result.}
}
\value{
A formula that can be executed in a dplyr summarize statement
}
\description{
Calculate a rate using a logical field, with TRUE as the numerator, and TRUE/FALSE
as the denominator. This excludes NA values. When provided, the results of each
are multiplied by the cases parameter, which allows us to appropriately scale 
the calculations for our reduced record data sets that are included in the package'
}
\details{
This function returns a function, meant to be applied during a dplyr summarize
statement, although it can be used in a broader context.
}
\examples{
 library(dplyr)
 births \%>\% 
     filter(DOB_YY >= 1989) \%>\% 
     group_by(DOB_YY) \%>\%
     summarize(
         cesarean_rate = lg_rate(cesarean_lg, cases)()
     ) \%>\%
     plot
}

