% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_aggregators.R
\name{wtd_lg_rate}
\alias{wtd_lg_rate}
\title{Weighted Logical Rate Calculation}
\usage{
wtd_lg_rate(logical_column, weight_column = "cases", na.rm = FALSE)
}
\arguments{
\item{logical_column}{a name of a logical table field name that will be aggregated}

\item{weight_column}{a value that should be applied to the result in order to simulate}

\item{na.rm}{a logical indicator of whether NA values should be removed prior to aggregation
a number of records with this result.}
}
\value{
a formula that can be executed in a dplyr summarize statement
}
\description{
Calculate a rate using a logical field, with TRUE as the numerator, and TRUE/FALSE as the denominator. This excludes NA values. When provided, the results of each are multiplied by the cases parameter, which allows us to appropriately scale the calculations for our reduced record data sets that are included in the package'
}
