% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{births}
\alias{births}
\title{US Birth Certificate}
\format{An object of class \code{\link{data.frame}}}
\source{
CDC and this package
}
\usage{
births
}
\description{
The primary analytic data set that is included in this package. This data frame
is the result of the raw data processing that is applied to birth certificate
data sets provided by the CDC.
}
\details{
\strong{pseudo_birth_date}: a date field which represents the year and month
of birth, with the day of the month fixed to the first day of the month. This
is due to the fact that we cannot be any more precise with this public data set.
We include this along with factor representations of the birth year and month
as separate fields, since it can be useful to have an actual date field value.

\strong{birth_year}: an ordered factor representing the year of birth.

\strong{birth_month}: an ordered factor representing the month of the birth.

\strong{birth_state}: a factor representing the state where the birth occured.
This dimension stopped being reported by the CDC after 2004 due to privacy concerns,
and all records after that point are marked as NA.

\strong{birth_in_hospital}: a logical representation of whether the birth occured
in a hospital. Missing records are represented by NA.

\strong{cesarean}: a logical representation of whether the birth occured via
cesarean section. In cases where the method of birth is unknown, the record is
represented with an NA. Birth method is not reported in records prior to 1989,
so all earlier records are represented as NA.

\strong{cases}: an integer representing the number of birth records that are
represented by the combination of dimensions that are present in a particular
record of the births data set. All math that is performed on this data set should
be weighted by this value.
}
\examples{
# Load data directly from the vitalstatistics namespace
vitalstatistics::births

# Or first load the entire package
library(vitalstatistics)
births

}
\keyword{datasets}

