% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{births}
\alias{births}
\title{US Birth Certificate}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 7818974 rows and 9 columns.}
\source{
CDC and this package
}
\usage{
births
}
\value{
an object of class \code{\link[data.table]{data.table}}
}
\description{
This is the primary analytic data set that included and used by this package. This data frame is
the result of the raw data processing that is applied to birth certificate data sets provided by
the CDC.
}
\details{
We include a description of each field in this data set, since they are the result of a quite
complicated process which blends data from over 40 years of distinct data sets with different
field values and usages.
}
\section{Conceptual Prefixes}{


Each field is named with a conceptual prefix in order to make it more clear. This is useful to
delineate fields that might be confusing, such as date of birth. If we were to simply use "date"
as our field name, it leaves some question as to whether this is the date of birth, date of
conception, date of reporting, etc. Instead, we use one of a defined set of words as the first
word in every field to make the meaning clear.

\strong{birth}: fields prefixed with "birth_" describe values relative to the delivery event. For
example, the "birth_hour" is intended to be the time when the delivery was completed, and
"birth_in_hospital" refers to whether the final moment of the delivery occured in the hospital,
regardless of the length of time that might have been spent in labor outside the hospital.

\strong{mother}: the mother is the primary subject addressed by this data set. Unless otherwise
stated in the field definition, all dependent attributes of the mother should be considered in
the context of the birth. For example, the age of the mother calculated at the time of the birth,
her state of residence at the time of the birth, and so forth.

\strong{child}: the child (i.e. newly born infant) is the secondary subject addressed by this
data set.
}

\section{Fields}{


  \strong{birth_in_hospital}: a logical representation of whether the birth occurred in a
  hospital. Missing records are represented by NA.

  \strong{birth_via_cesarean}: a logical representation of whether the birth occured via cesarean
  section. In cases where the method of birth is unknown, the record  is represented with an NA.
  Birth method is not reported in records prior to 1989, so all earlier records are represented
  as NA.

  \strong{mother_age_int}: an integer value which represents the age (in years) of the mother at
  the time of delivery.

  \strong{mother_age}: a factor representation of the age (in years) of the mother at the time of
  delivery. This field differs from mother_age_int in that it uses banded values for ages 10-12
  and 50-54. The tradeoff for the loss of precision is that this field is slightly better
  populated than mother_age_int.

  \strong{child_sex}: a factor representation of the sex of the infant.

  \strong{cases}: an integer representing the number of birth records that are represented by the
  combination of dimensions that are present in a particular record of the births data set. All
  math that is performed on this data set should be weighted by this value.
}

\keyword{datasets}
