% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{featherize}
\alias{featherize}
\title{Vital Statistics Data into Feathers}
\usage{
featherize()
}
\description{
Exports all data sets included in this package into the feather format. This is intended for use
by the Python pandas package.
}
