% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_aggregators.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Numeric Value Summary of Weighted Records}
\usage{
numeric_summary(data, numeric_column, weight_column = "cases",
  na.rm = FALSE)
}
\arguments{
\item{data}{a data frame, presumably the births data set or a derivative}

\item{numeric_column}{the numeric column that you want to perform summary statistics on}

\item{weight_column}{the column in the data.frame that contains the weighting value}

\item{na.rm}{whether to pass a TRUE or FALSE value to the na.rm argument for each underlying aggregation function.}
}
\value{
A formula that can be executed in a \code{\link{dplyr}} summarize statement
}
\description{
Because the \code{\link{births}} data set uses a weighted record strategy (i.e. you have to multiply everything by the cases field), the typical summary function won't return meaningful results. In order to provide some basic descriptive statistics for a numeric column in the data set, this function can be used instead.
}
\details{
It makes use of the dplyr format for summarizing results, and therefore integrates nicely with a chain of dplyr functions. Under the hood, it is using \code{\link[dplyr]{summarize_}} and pasting strings together for evaluation, with the actual statistics being handled by the \code{\link{matrixStats}} package, based upon your input.
}
